==> notebooks/00_get_data.ipynb <==
# Get the vehicles dataset

import kagglehub
import pandas as pd
import os

data_path = kagglehub.dataset_download("nehalbirla/vehicle-dataset-from-cardekho")

csv_path = os.path.join(data_path, "Car details v3.csv")

df = pd.read_csv(csv_path)
df.head()

from pathlib import Path


to_path = Path("../datasets/cars.csv")
to_path.parent.mkdir(parents=True, exist_ok=True)

df.to_csv(to_path, index=False)
