==> notebooks/06_ml_pipeline.ipynb <==
# ML Pipeline

from pathlib import Path

import pandas as pd
import numpy as np

from sklearn.pipeline import Pipeline, FunctionTransformer
from sklearn.impute import SimpleImputer
from sklearn.preprocessing import StandardScaler, OneHotEncoder
from sklearn.compose import ColumnTransformer

from sklearn import set_config
set_config(transform_output="pandas")

train_path = Path("../datasets/train.csv")

cars = pd.read_csv(train_path)

cars.head()

cars.info()

cars_labels = cars['selling_price'].copy()
cars = cars.drop('selling_price', axis=1)

cars_labels.head()

cars.head()

## Preprocessing Pipeline

### Helper functions

def normalize_mileage(df: pd.DataFrame) -> pd.Series:
    df = df.copy()

    df['mileage_unit'] = df['mileage'].str.split().str[-1]
    df['mileage'] = df['mileage'].astype(str).str.extract(r'([\d\.]+)', expand=False)
    df['mileage'] = pd.to_numeric(df['mileage'], errors='coerce')

    mask_petrol = (df['mileage_unit'] == "km/kg") & (df['fuel'] == "Petrol")
    mask_diesel = (df['mileage_unit'] == "km/kg") & (df['fuel'] == "Diesel")
    mask_cng = (df['mileage_unit'] == "km/kg") & (df['fuel'] == "CNG")
    mask_lpg = (df['mileage_unit'] == "km/kg") & (df['fuel'] == "LPG")

    df.loc[mask_petrol, 'mileage'] /= 0.74
    df.loc[mask_diesel, 'mileage'] /= 0.832
    df.loc[mask_lpg, 'mileage'] /=   0.54
    df.loc[mask_cng, 'mileage'] /=   0.128

    return df['mileage']

def group_seats(df : pd.DataFrame) -> pd.DataFrame:
    df = df.copy()

    conditions = [
        df['seats'] < 5,
        df['seats'] == 5,
        df['seats'] > 5
    ]
    choices = ['less_than_five', 'five', 'more_than_five']
    # ensure default has the same dtype (string) as choices to avoid dtype promotion errors
    df['seats'] = np.select(conditions, choices, default='missing')
    df['seats'] = df['seats'].astype('category')

    return df


from category_encoders import TargetEncoder

target_encoder = TargetEncoder(
    cols=["brand_effect"],
    smoothing=10,
    handle_unknown="value",
    handle_missing="value"
)

def group_rare_names(df: pd.DataFrame, threshold=10) -> pd.Series:
    df = df.copy()

    name_counts = df['name'].value_counts()
    rare_names = name_counts[name_counts < threshold].index
    df['name'] = df['name'].replace(rare_names, 'other')

    return df['name']

def group_rare_fuel(df : pd.DataFrame) -> pd.DataFrame:
    df = df.copy()
    df['fuel'] = df['fuel'].replace({
        'CNG': 'other',
        'LPG': 'other'
    })
    return df

def update_owner_grouping(df: pd.DataFrame) -> pd.DataFrame:
    df = df.copy()
    df['owner'] = df['owner'].replace({
        'Third Owner': 'Third & Above Owner',
        'Fourth & Above Owner': 'Third & Above Owner',
        'Test Drive Car': 'First Owner'
    })
    return df

from sklearn.preprocessing import OrdinalEncoder


ordinal_encoder = OrdinalEncoder(
    handle_unknown="use_encoded_value",
    unknown_value=-1,
    categories=[["Third & Above Owner", "Second Owner", "First Owner"]]
)

def convert_year_to_age(df: pd.DataFrame) -> pd.Series:
    df = df.copy()
    df['age'] = 2026 - df['year']
    return df['age']

### Custom Transformers

from sklearn.base import BaseEstimator, TransformerMixin


class BaseNumericFeatures(BaseEstimator, TransformerMixin):
    def fit(self, X, y=None):
        return self

    def transform(self, X):
        X = X.copy()

        X["engine"] = pd.to_numeric(
            X["engine"].str.split().str[0],
            errors="coerce"
        )

        X["max_power"] = pd.to_numeric(
            X["max_power"].str.split().str[0],
            errors="coerce"
        )

        X["mileage"] = normalize_mileage(X)
        X["age"] = convert_year_to_age(X)

        X.drop(columns=['year', 'fuel'], inplace=True)

        return X

    def get_feature_names_out(self, input_features=None):
        return ["engine", "max_power", "mileage", "age", "km_driven"]

class InteractionFeatures(BaseEstimator, TransformerMixin):
    def fit(self, X, y=None):
        return self

    def transform(self, X):
        eps = 1e-6
        X = X.copy()

        X["engine_mileage_ratio"] = X["engine"] / (X["mileage"] + eps)
        X["km_driven_age_interaction"] = X["km_driven"] * X["age"]

        return X

    def get_feature_names_out(self, input_features=None):
        return [
                "engine_mileage_ratio",
                 "km_driven_age_interaction",
                   "engine", "max_power", "mileage", "age", "km_driven"]

class NameTranformation(BaseEstimator, TransformerMixin):
    def __init__(self, threshold=5):  # no *args or **kwargs!
        self.threshold = threshold

    def fit(self, X, y=None):
        return self

    def transform(self, X):
        X = X.copy()

        X['name'] = X['name'].str.split().str[0]
        X['name'] = group_rare_names(X, self.threshold)

        X.rename(columns={"name": "brand_effect"}, inplace=True)
        
        return X

    def get_feature_names_out(self, input_features=None):
        return ["brand_effect"]

### Pipelines

num_inter_pip = Pipeline([
    ("base_numeric_features", BaseNumericFeatures()),
    ("interaction_features", InteractionFeatures()),
    ("median_imputer", SimpleImputer(strategy='median')),
    ("log", FunctionTransformer(np.log1p, feature_names_out='one-to-one')),
    ("standarize", StandardScaler()),
])

seats_pip = Pipeline([
    ("median_imputer", SimpleImputer(strategy='median')),
    ("regroup", FunctionTransformer(group_seats, feature_names_out='one-to-one')),
    ("one-hot encoding", OneHotEncoder(sparse_output=False, handle_unknown='ignore'))
])

name_pip = Pipeline([
    ("regroup", NameTranformation()),
    ("target_encode", target_encoder),
    ("standarize", StandardScaler()),
])

fuel_pip = Pipeline([
    ("group_rare", FunctionTransformer(group_rare_fuel, feature_names_out='one-to-one')),
    ("one-hot encoding", OneHotEncoder(sparse_output=False, handle_unknown='ignore'))
])

one_hot_pip = Pipeline([
    ("encoding", OneHotEncoder(sparse_output=False, handle_unknown='ignore'))
])

owner_pip = Pipeline([
    ("regroup", FunctionTransformer(update_owner_grouping, feature_names_out='one-to-one')),
    ("ordinal_encoding", ordinal_encoder)
])

### Preprocessing Column Transformer

preproc = ColumnTransformer([
    ("num_inter", num_inter_pip, ['engine', 'max_power', 'fuel', 'year', 'mileage', 'km_driven']),
    ("seats", seats_pip, ['seats']),
    ("name", name_pip, ['name']),
    ("fuel", fuel_pip, ['fuel']),
    ("one_hot_encoding", one_hot_pip, ['seller_type', 'transmission']),
    ("owner", owner_pip, ['owner']),
    
])
# preproc.fit_transform(cars)

## Full ML Pipeline

from sklearn.linear_model import Ridge

pipe = Pipeline([
    ("preproc", preproc),
    ("model", Ridge())
])

from sklearn.compose import TransformedTargetRegressor


full_pipe = TransformedTargetRegressor(
    regressor=pipe,
    func=np.log1p,
    inverse_func=np.expm1
)

full_pipe.fit(cars, cars_labels)

cars_predictions = full_pipe.predict(cars)
print(f" first seven pridictions: {cars_predictions[:7].round(2)}")
print(f" first seven real valuse: {cars_labels.iloc[:7].values}")

from sklearn.metrics import root_mean_squared_error


ridge_rmse = root_mean_squared_error(cars_labels, cars_predictions)
ridge_rmse

import matplotlib.pyplot as plt

residuals = cars_labels - cars_predictions

plt.scatter(cars_predictions, residuals, alpha=0.3)
plt.axhline(0, color="black")
plt.xlabel("Predicted")
plt.ylabel("Residuals")
plt.title("Residuals vs Predictions")
plt.show()

### Evaluate With Cross Validation

from sklearn.metrics import mean_absolute_error, root_mean_squared_error, r2_score, mean_absolute_percentage_error
from sklearn.model_selection import KFold, cross_validate

scoring = {
    "mae": "neg_mean_absolute_error",
    "mape": "neg_mean_absolute_percentage_error",
    "rmse": "neg_root_mean_squared_error",
    "r2": "r2"
}

cv = KFold(
    n_splits=10,
    shuffle=True,
    random_state=42
)

def print_cv_summery(cv_results):
    cv_df = pd.DataFrame(cv_results)

    summary = pd.DataFrame({
        "metric": ["MAE", "MAPE", "RMSE", "R2"],
        "mean": [
            -cv_df["test_mae"].mean(),
            -cv_df["test_mape"].mean(),
            -cv_df["test_rmse"].mean(),
            cv_df["test_r2"].mean()
        ],
        "std": [
            cv_df["test_mae"].std(),
            cv_df["test_mape"].std(),
            cv_df["test_rmse"].std(),
            cv_df["test_r2"].std()
        ]
    })

    print(summary)

#### Ridge Model

cv_results = cross_validate(
    estimator=pipe,
    X=cars,
    y=cars_labels,
    scoring=scoring,
    cv=cv,
    return_train_score=True,
    n_jobs=-1
)

print_cv_summery(cv_results)

#### Decision Tree Regressor

from sklearn.tree import DecisionTreeRegressor

pipe = Pipeline([
    ("preproc", preproc),
    ("model", DecisionTreeRegressor(random_state=42))
])

full_pipe = TransformedTargetRegressor(
    regressor=pipe,
    func=np.log1p,
    inverse_func=np.expm1
)

cv_results = cross_validate(
    estimator=pipe,
    X=cars,
    y=cars_labels,
    scoring=scoring,
    cv=cv,
    return_train_score=True,
    n_jobs=-1
)

print_cv_summery(cv_results)

#### Random Forest Regressor

from sklearn.ensemble import RandomForestRegressor

pipe = Pipeline([
    ("preproc", preproc),
    ("model", RandomForestRegressor(random_state=42))
])

full_pipe = TransformedTargetRegressor(
    regressor=pipe,
    func=np.log1p,
    inverse_func=np.expm1
)

cv_results = cross_validate(
    estimator=pipe,
    X=cars,
    y=cars_labels,
    scoring=scoring,
    cv=cv,
    return_train_score=True,
    n_jobs=-1
)

print_cv_summery(cv_results)

full_pipe.fit(cars, cars_labels)
cars_predictions = full_pipe.predict(cars)
forest_rmse = root_mean_squared_error(cars_labels, cars_predictions)
forest_rmse

full_pipe.fit(cars, cars_labels)
cars_predictions = full_pipe.predict(cars)
forest_mae = mean_absolute_error(cars_labels, cars_predictions)
forest_mae

forest_mape = mean_absolute_percentage_error(cars_labels, cars_predictions)
forest_mape

forest_r2 = r2_score(cars_labels, cars_predictions)
forest_r2

import matplotlib.pyplot as plt

residuals = cars_labels - cars_predictions

plt.scatter(cars_predictions, residuals, alpha=0.3)
plt.axhline(0, color="black")
plt.xlabel("Predicted")
plt.ylabel("Residuals")
plt.title("Residuals vs Predictions")
plt.show()

plt.scatter(cars_predictions, cars_labels, alpha=0.3)
plt.axhline(0, color="black")
plt.xlabel("Predicted")
plt.ylabel("labels")
plt.title("labels vs Predictions")
plt.show()

plt.hist(residuals, bins=50)
plt.title("Residual Distribution")
plt.show()

#### Notes:
- Random Forest was the best model with CV
- When we train the RF model on CV its perfromance was much less than on training => there is overfitting

### Fine Tune my RF model

from scipy.stats import randint


pipe = Pipeline([
    ("preproc", preproc),
    ("random_forest", RandomForestRegressor())
])

full_pipe = TransformedTargetRegressor(
    regressor=pipe,
    func=np.log1p,
    inverse_func=np.expm1
)

param_dist = {
    "regressor__preproc__name__regroup__threshold": randint(1, 31),
    
    "regressor__random_forest__n_estimators": randint(50, 800),
    "regressor__random_forest__max_depth": [2, 4, 8, 16, 24, 32],
    "regressor__random_forest__min_samples_leaf": randint(2, 150),
    "regressor__random_forest__max_features": ["sqrt", "log2", 0.3, 0.6, 0.8]
}

cv = KFold(
    n_splits=10,
    shuffle=True,
    random_state=42
)

from sklearn.model_selection import RandomizedSearchCV


search = RandomizedSearchCV(
    estimator=full_pipe,
    param_distributions=param_dist,
    n_iter=200,
    scoring='neg_root_mean_squared_error',
    cv=cv,
    n_jobs=-1,
    verbose=2,
    random_state=42,
    return_train_score=True,

)

search.fit(cars, cars_labels)

-search.best_score_

cv_res = pd.DataFrame(search.cv_results_)
cv_res.sort_values(by="mean_test_score", ascending=False, inplace=True)
# rename columns instead of replace (replace operates on values)
cv_res.rename(columns={
    'param_regressor__random_forest__n_estimators': 'rf_n_estimators',
    'param_regressor__random_forest__max_depth': 'rf_max_depth',
    'param_regressor__random_forest__min_samples_leaf': 'rf_min_samples_leaf',
    'param_regressor__random_forest__max_features': 'rf_max_features',
}, inplace=True)
cv_res.head()

search.best_params_

final_model = search.best_estimator_  # includes preprocessing
feature_importances = final_model.regressor_["random_forest"].feature_importances_

sorted(zip(feature_importances,
           final_model.regressor_["preproc"].get_feature_names_out()),
       reverse=True)

final_model.fit(cars, cars_labels)

cars_predictions = final_model.predict(cars)

print(cars_labels.head())
print(cars_predictions[:5])

root_mean_squared_error(cars_labels, cars_predictions)

r2_score(cars_labels, cars_predictions)

mean_absolute_error(cars_labels, cars_predictions)

plt.style.use('dark_background')

residuals = cars_labels - cars_predictions

plt.scatter(cars_predictions, residuals, alpha=0.3)
plt.axhline(0, color="black")
plt.xlabel("Predicted")
plt.ylabel("Residuals")
plt.title("Residuals vs Predictions")
plt.show()

plt.scatter(cars_predictions, cars_labels, alpha=0.3)
plt.axhline(0, color="black")
plt.xlabel("Predicted")
plt.ylabel("labels")
plt.title("labels vs Predictions")
plt.show()

plt.hist(residuals, bins=50)
plt.title("Residual Distribution")
plt.show()

### Metrices on train and validation set

We can see that:
- milage, km_driven_age_interaction, engine_mileage_ratio, brand_effect, km_driven, and age are **the most important features**
- rmse_validation = 162935.01
- r2_train = 0.97 
- rmse_train = 137623.25
- mae_train = 55700.84

### Save model to .pkl file

import joblib
joblib.dump(final_model, Path("../models/rf_model_pipeline_v1.pkl"))

### Ploting the learning curve

from sklearn.model_selection import learning_curve
from sklearn.compose import TransformedTargetRegressor


import matplotlib.pyplot as plt

pipe = Pipeline([
    ("preproc", preproc),
    ("model", RandomForestRegressor(
        n_estimators=65,
        max_features=0.8,
        max_depth=24,
        min_samples_leaf=5
    ))
])

full_pipe = TransformedTargetRegressor(
    regressor=pipe,
    func=np.log1p,
    inverse_func=np.expm1
)



train_sizes, train_scores, valid_scores = learning_curve(
    full_pipe,
    cars,
    cars_labels,
    train_sizes=np.linspace(0.01, 1.0, 40),
    cv=10,
    scoring="neg_root_mean_squared_error"
    )

train_errors = -train_scores.mean(axis=1)
valid_errors = -valid_scores.mean(axis=1)


plt.figure(figsize=(15,12))
plt.plot(train_sizes, train_errors, "r-+", linewidth=2, label="train")
plt.plot(train_sizes, valid_errors, "b-", linewidth=3, label="valid")
plt.legend(loc="upper right")
plt.xlabel("Training set size")
plt.ylabel("RMSE")
plt.grid()
plt.axis([0, 2e4, 0, 1e6])

plt.show()

## Test the model

test_path = Path("../datasets/test.csv")

cars_test = pd.read_csv(test_path)

cars_test_labels = cars_test['selling_price'].copy()
cars_test = cars_test.drop('selling_price', axis=1)

import joblib
from pathlib import Path


rf_model = joblib.load(Path("../models/rf_model_pipeline_v1.pkl"))
rf_model

test_predictions = rf_model.predict(cars_test)

cars_test.head(1)

cars_test.info()

print(f"Test RMSE: {root_mean_squared_error(cars_test_labels, test_predictions):.2f}")
print(f"r2: {r2_score(cars_test_labels, test_predictions):.4f}")
print(f"mae: {mean_absolute_error(cars_test_labels, test_predictions):.2f}")

### Metrices on test set

- rmse: 148360.72
- r2 = 0.96
- mae: 71207.18
